NIFI-6729 - AzureStorageCredentialsControllerServiceLookup no longer implements LookupService. Also moved it from org.apache.nifi.lookup to org.apache.nifi.service.lookup.
